Refactor sliding nav to match vanilla design
