+++ templates/templates/_navigation.html
@@ -9,8 +9,6 @@
                     <a class="p-navigation__link" href="/">
                         {% if breadcrumbs and breadcrumbs.children or breadcrumbs.grandchildren %}
                             Canonical Ubuntu
-                        {% elif breadcrumbs == "tutorials" %}
-                            Canonical Ubuntu
                         {% else %}
                             <div class="p-navigation__logo-tag">
                                 <img class="p-navigation__logo-icon"
@@ -28,7 +26,7 @@
                            aria-label="Search"></a>
                     </li>
                     <li class="p-navigation__item">
-                        <a href="/navigation" class="js-menu-button p-navigation__link">Menu</a>
+                        <a href="#" class="js-menu-button p-navigation__link">Menu</a>
                     </li>
                 </ul>
             </div>
@@ -84,6 +82,7 @@
                            tabindex="0"
                            onfocus="fetchDropdown('/templates/meganav/get-ubuntu', 'get-ubuntu');">Get Ubuntu</a>
                     </li>
+                    
                     <li class="p-navigation__item--dropdown-toggle global-nav-mobile global-nav"
                         role="menuitem"
                         id="all-canonical"></li>
