+++ templates/templates/base.html
@@ -24,7 +24,6 @@
         <script src="https://assets.ubuntu.com/v1/703e23c9-lazysizes+noscript+native-loading.5.1.2.min.js"
                 defer></script>
 
-        <script src="{{ versioned_static('js/src/navigation.js') }}" defer></script>
         <script src="{{ versioned_static('js/dist/main.js') }}" defer></script>
         <script src="{{ versioned_static('js/src/infer-preferred-language.js')}}"></script>
